+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * add new WP-Cron task to run concurrently to `wp_version_check` so that Git Updater managed plugins and themes can take advantage of auto updating
+* added better check to see if background updating cron event is already scheduled
 
 #### 10.3.4 / 2021-06-22
 * refactor `get_repo_requirements()`